asm/c64: Adding asm stubs for dot functions.
